more code clean up change package to io.nsq
