+++ pom.xml
@@ -4,9 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.github.dustismo</groupId>
-  <artifactId>trendrr-nsq-client</artifactId>
-  <version>1.2-SNAPSHOT</version>
+  <groupId>io.nsq</groupId>
+  <artifactId>client</artifactId>
+  <version>1.0.0-SNAPSHOT</version>
 
   <dependencies>
     <dependency>
@@ -24,7 +24,7 @@
     <dependency>
       <groupId>io.netty</groupId>
       <artifactId>netty-all</artifactId>
-      <version>4.0.25.Final</version>
+      <version>4.0.26.Final</version>
     </dependency>
 
     <dependency>
@@ -38,6 +38,12 @@
       <artifactId>jackson-databind</artifactId>
       <version>2.5.1</version>
     </dependency>
-  </dependencies>
 
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>4.12</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
 </project>
