--- src/main/java/com/trendrr/nsq/NSQMessageCallback.java
@@ -1,8 +1,0 @@
-package com.trendrr.nsq;
-
-public interface NSQMessageCallback {
-
-	public void message(NSQMessage message);
-
-	public void error(Exception x);
-}
