--- src/main/java/com/trendrr/nsq/exceptions/NoConnectionsException.java
@@ -1,9 +1,0 @@
-package com.trendrr.nsq.exceptions;
-
-public class NoConnectionsException extends Exception {
-	private static final long serialVersionUID = -7902210058291470928L;
-
-	public NoConnectionsException(String message, Throwable cause) {
-		super(message, cause);
-	}
-}
