+++ src/main/java/io/nsq/ConnectionAddress.java
@@ -1,0 +1,33 @@
+package io.nsq;
+
+public class ConnectionAddress {
+    private int poolsize = 1;
+
+    public int getPoolsize() {
+        return poolsize;
+    }
+
+    public void setPoolsize(int poolsize) {
+        this.poolsize = poolsize;
+    }
+
+
+    public String getHost() {
+        return host;
+    }
+
+    public void setHost(String host) {
+        this.host = host;
+    }
+
+    public int getPort() {
+        return port;
+    }
+
+    public void setPort(int port) {
+        this.port = port;
+    }
+
+    private String host;
+    private int port;
+}
