+++ src/main/java/io/nsq/Connections.java
@@ -1,0 +1,104 @@
+package io.nsq;
+
+import com.google.common.collect.Maps;
+import io.nsq.exceptions.NoConnectionsException;
+import org.apache.logging.log4j.LogManager;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+
+public class Connections {
+	protected HashMap<String, List<Connection>> connections = Maps.newHashMap();
+	protected List<Connection> connectionList = new ArrayList<>();
+	protected long rrCount = 0;
+	
+	public synchronized void addConnection(Connection connection) {
+		if (connection == null) {
+			return;
+		}
+		
+		String key = connection.getHost() + ":" + connection.getPort();
+		if (connectionList.contains(connection)) {
+            LogManager.getLogger(this).warn("Already contains connection:" + connection);
+			return;
+		}
+		
+		if (connections.get(key) == null) {
+			connections.put(key, new ArrayList<>());
+		}
+		connections.get(key).add(connection);
+		connectionList.add(connection);
+	}
+	
+	public synchronized boolean containsConnection(String host, int port) {
+		return connections.containsKey(this.getKey(host, port));
+	}
+
+    public synchronized int connectionSize(String host, int port) {
+		String key = this.getKey(host, port);
+		List<Connection> c = this.connections.get(key);
+		if (c == null)
+			return 0;
+		return c.size();
+	}
+	
+	public synchronized int size() {
+		return this.connectionList.size();
+	}
+
+    public synchronized List<Connection> getConnections() throws NoConnectionsException {
+		//return a copy of the internal list so we dont have any
+		//concurrent modification problems.
+		return new ArrayList<Connection>(this.connectionList);
+	}
+
+    public synchronized Connection next() throws NoConnectionsException {
+		if (connectionList.size() == 0) {
+            LogManager.getLogger(this).warn("No connections available!");
+			throw new NoConnectionsException("No connections available", null);
+		}
+		
+		return connectionList.get((int)(rrCount++ % connectionList.size()));
+	}
+	
+	public synchronized void remove(Connection connection) {
+		String key = connection.getHost() + ":" + connection.getPort();
+		if (!connectionList.contains(connection)) {
+            LogManager.getLogger(this).warn("Does not contain : " + key + ": " + connection);
+			return;
+		}
+		List<Connection> conn = connections.get(key);
+		conn.remove(connection);
+		if (conn.isEmpty()) {
+			connections.remove(key);
+		}
+		this.connectionList.remove(connection);
+        LogManager.getLogger(this).warn("REMOVED: " + key);
+	}
+	
+	public synchronized void remove(String host, int port) {
+		List<Connection> conns = this.connections.remove(host + ":" + port);
+		if (conns == null) {
+            LogManager.getLogger(this).warn("no connection : " + host + ":" + port);
+			return;
+		}
+		this.connectionList.removeAll(conns);		
+	}
+	
+	public synchronized void close() {
+		try {
+			for (Connection c: this.getConnections()) {
+				c.close();
+			}
+		} catch (NoConnectionsException e) {
+//			log.error("Caught", e);
+		}
+		this.connections.clear();
+		this.connectionList.clear();
+	}
+	
+	public String getKey(String host, int port) {
+		return host + ":" + port;
+	}
+}
