+++ src/main/java/io/nsq/callbacks/BatchSuccessCallback.java
@@ -1,0 +1,7 @@
+package io.nsq.callbacks;
+
+@FunctionalInterface
+public interface BatchSuccessCallback {
+
+    public void batchSuccess(String topic, int num);
+}
