+++ src/main/java/io/nsq/exceptions/BadMessageException.java
@@ -1,0 +1,8 @@
+package io.nsq.exceptions;
+
+public class BadMessageException extends NSQException {
+
+	public BadMessageException(String message) {
+		super(message);
+	}
+}
