+++ src/main/java/io/nsq/exceptions/BadTopicException.java
@@ -1,0 +1,8 @@
+package io.nsq.exceptions;
+
+public class BadTopicException extends NSQException {
+
+	public BadTopicException(String message) {
+		super(message);
+	}
+}
