+++ src/main/java/io/nsq/exceptions/DisconnectedException.java
@@ -1,0 +1,8 @@
+package io.nsq.exceptions;
+
+public class DisconnectedException extends NSQException {
+
+	public DisconnectedException(String message, Throwable cause) {
+		super(message, cause);
+	}
+}
