+++ src/main/java/io/nsq/exceptions/NoConnectionsException.java
@@ -1,0 +1,8 @@
+package io.nsq.exceptions;
+
+public class NoConnectionsException extends NSQException {
+
+	public NoConnectionsException(String message, Throwable cause) {
+		super(message, cause);
+	}
+}
