+++ src/main/java/io/nsq/netty/NSQClientInitializer.java
@@ -1,0 +1,16 @@
+package io.nsq.netty;
+
+import io.netty.channel.ChannelInitializer;
+import io.netty.channel.ChannelPipeline;
+import io.netty.channel.socket.SocketChannel;
+
+public class NSQClientInitializer extends ChannelInitializer<SocketChannel> {
+    @Override
+    protected void initChannel(SocketChannel ch) throws Exception {
+        ChannelPipeline pipeline = ch.pipeline();
+
+        pipeline.addLast(new NSQDecoder());
+        pipeline.addLast(new NSQEncoder());
+        pipeline.addLast(new NSQHandler());
+    }
+}
